Tile: Lowering dispatch, resident-accumulator naming, tile/ split
- 'Hoist' is gone: the concept is a *resident accumulator* (the classic
  init_accumulator / epilogue) and the whole lifecycle is one entry
  point, mma_resident, in matmul/resident.rs (was acc.rs). lower.rs no
  longer sees init/drain.
- Tile::mma dispatches on one comptime classification: Lowering::of
  (Leaf | Resident | Partition | Walk(schedule)) from Tile::class()
  (TileClass replaces the ad-hoc is_mem bool soup).
- tile.rs (1142 lines) split by backing store: tile/mod.rs holds the
  Tile surface and kind-dispatching ops; mem.rs, cmma.rs, tma.rs hold
  each variant's data and leaves.
