+++ crates/cubek-tile/src/lib.rs
@@ -9,7 +9,6 @@ mod quantization;
 mod space;
 mod stage;
 mod tile;
-mod tile_kind;
 mod view;
 
 // `Axis`/`MAX_AXES` and `ConcreteLayout` are the storage-layout vocabulary; clients reach them
@@ -21,5 +20,4 @@ pub use quantization::*;
 pub use space::*;
 pub use stage::*;
 pub use tile::*;
-pub use tile_kind::*;
 pub use view::*;
